Remove dependency to logging module in JDBC
