+++ mode/core/pom.xml
@@ -33,11 +33,6 @@
             <artifactId>shardingsphere-infra-context</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>ch.qos.logback</groupId>
-            <artifactId>logback-classic</artifactId>
-            <scope>compile</scope>
-        </dependency>
         
         <dependency>
             <groupId>org.apache.shardingsphere</groupId>
