+++ mode/core/src/main/java/org/apache/shardingsphere/mode/metadata/MetaDataContexts.java
@@ -17,11 +17,7 @@
 
 package org.apache.shardingsphere.mode.metadata;
 
-import ch.qos.logback.classic.Level;
-import ch.qos.logback.classic.Logger;
-import ch.qos.logback.classic.LoggerContext;
 import lombok.Getter;
-import org.apache.shardingsphere.infra.config.props.ConfigurationPropertyKey;
 import org.apache.shardingsphere.infra.database.type.dialect.OpenGaussDatabaseType;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
 import org.apache.shardingsphere.infra.metadata.data.ShardingSphereData;
@@ -32,11 +28,9 @@ import org.apache.shardingsphere.infra.metadata.data.builder.ShardingSphereDataB
 import org.apache.shardingsphere.infra.rule.identifier.type.ResourceHeldRule;
 import org.apache.shardingsphere.infra.util.spi.type.typed.TypedSPILoader;
 import org.apache.shardingsphere.mode.metadata.persist.MetaDataPersistService;
-import org.slf4j.LoggerFactory;
 
 import java.util.Map.Entry;
 import java.util.Optional;
-import java.util.Properties;
 
 /**
  * Meta data contexts.
@@ -68,8 +62,6 @@ public final class MetaDataContexts implements AutoCloseable {
         Optional<ShardingSphereData> loadedShardingSphereData = Optional.ofNullable(persistService.getShardingSphereDataPersistService())
                 .flatMap(shardingSphereDataPersistService -> shardingSphereDataPersistService.load(metaData));
         loadedShardingSphereData.ifPresent(optional -> useLoadedToReplaceInit(result, optional));
-        // TODO skip this while instance type is JDBC
-        refreshRootLogger(metaData.getProps().getProps());
         return result;
     }
     
@@ -95,16 +87,6 @@ public final class MetaDataContexts implements AutoCloseable {
                 entry.setValue(loadedSchemaData.getTableData().get(entry.getKey()));
             }
         }
-    }
-    
-    private void refreshRootLogger(final Properties props) {
-        LoggerContext loggerContext = (LoggerContext) LoggerFactory.getILoggerFactory();
-        Logger rootLogger = loggerContext.getLogger(Logger.ROOT_LOGGER_NAME);
-        renewRootLoggerLevel(rootLogger, props);
-    }
-    
-    private void renewRootLoggerLevel(final Logger rootLogger, final Properties props) {
-        rootLogger.setLevel(Level.valueOf(props.getOrDefault(ConfigurationPropertyKey.SYSTEM_LOG_LEVEL.getKey(), ConfigurationPropertyKey.SYSTEM_LOG_LEVEL.getDefaultValue()).toString()));
     }
     
     @Override
